Move hyrax_helper methods to new hyc_helper file and add sort_people_by_index method
