+++ app/controllers/application_controller.rb
@@ -2,6 +2,7 @@ class ApplicationController < ActionController::Base
   helper Openseadragon::OpenseadragonHelper
   # Adds a few additional behaviors into the application controller
   include HyraxHelper
+  include HycHelper
   include Blacklight::Controller
   include Hydra::Controller::ControllerBehavior
