+++ app/services/tasks/doi_create_service.rb
@@ -1,6 +1,6 @@
 module Tasks
   class DoiCreateService
-    include HyraxHelper
+    include HycHelper
 
     # From page 38 https://schema.datacite.org/meta/kernel-4.2/doc/DataCite-MetadataKernel_v4.2.pdf
     DCMI_TO_DATACITE_TYPE = {
